+++ CHANGES.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 
+- Fix ``merge`` and ``assign``/``extend`` so they apply ``clone_deep`` to source values before assigning to destination object.
 - Make ``merge`` accept a callback as a positional argument if it is last.